Add files via upload
java doc, conexión a bbdd separada, acceso por fichero propertis a los datos de conexion, .jar

+++ ProyectoPSP1EV/src/controllers/BBddControllers.java
@@ -1,11 +1,17 @@
 package controllers;
 
 import java.sql.Statement;
+import java.util.Properties;
 import java.util.Random;
+
+import javax.swing.JOptionPane;
 
 import models.NumerosAleatorios;
 
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.io.InputStream;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
@@ -40,7 +46,35 @@ public class BBddControllers {
 	}
 
 	/**
-	 * Método para insertar un registro en una BBDD mySQL
+	 * Conexión a la BBDD mySql
+	 * 
+	 * @return conexion
+	 */
+
+	// Conexión BBDD
+	public Connection conexionBbdd() {
+
+		Connection conexion = null;
+		Properties propiedades = new Properties();
+		InputStream entrada = null;
+		try {
+			entrada = new FileInputStream("datos.properties");
+			propiedades.load(entrada);
+			propiedades.getProperty("pass");
+			conexion = DriverManager.getConnection(propiedades.getProperty("servidor"), propiedades.getProperty("user"),
+					propiedades.getProperty("pass"));
+		} catch (SQLException e) {
+			System.out.println("Fallo en la conexión a la BBDD");
+		} catch (FileNotFoundException e) {
+			e.getLocalizedMessage();
+		} catch (IOException e) {
+			e.getLocalizedMessage();
+		}
+		return conexion;
+	}
+
+	/**
+	 * Método para insertar un registro en la BBDD mySQL
 	 * 
 	 * @param email
 	 * @param aleatorio
@@ -49,11 +83,14 @@ public class BBddControllers {
 	public void escribirBbdd(String email, int aleatorio) {
 
 		try {
-			Connection conexion = DriverManager.getConnection("jdbc:mysql://localhost/bbdd_psp_1", "DAM2020_PSP",
-					"DAM2020_PSP");
-			Statement consulta = conexion.createStatement();
-			consulta.executeUpdate("insert into empleados(EMAIL,INGRESOS) values('" + email + "'," + aleatorio + ")");
-			System.out.println(Thread.currentThread().getName() + " email: " + email + " salario: " + aleatorio);
+			Connection conexion = conexionBbdd();
+			if (conexion != null) {
+				Statement consulta = conexion.createStatement();
+				consulta.executeUpdate(
+						"insert into empleados(EMAIL,INGRESOS) values('" + email + "'," + aleatorio + ")");
+				System.out.println(Thread.currentThread().getName() + " email: " + email + " salario: " + aleatorio);
+				conexion.close();
+			}
 			conexion.close();
 		} catch (SQLException e) {
 			e.printStackTrace();
@@ -61,18 +98,19 @@ public class BBddControllers {
 	}
 
 	/**
-	 * Método para borrar todos registros de una BBDD mySQL
+	 * Método para borrar todos registros de la BBDD mySQL
 	 */
 
 	public void borrarBbdd() {
 
 		try {
-			Connection conexion = DriverManager.getConnection("jdbc:mysql://localhost/bbdd_psp_1", "DAM2020_PSP",
-					"DAM2020_PSP");
-			Statement consulta = conexion.createStatement();
-			int valor = consulta.executeUpdate("TRUNCATE TABLE empleados");
-			System.out.println("Tabla borrada!!");
-			conexion.close();
+			Connection conexion = conexionBbdd();
+			if (conexion != null) {
+				Statement consulta = conexion.createStatement();
+				int valor = consulta.executeUpdate("TRUNCATE TABLE empleados");
+				System.out.println("Tabla borrada!!");
+				conexion.close();
+			}
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
